requirements.txt
anaconda-anon-usage
annotated-types
archspec
asttokens
attrs==25.3.0
boltons
boxsdk==3.14.0
Brotli
certifi
cffi
charset-normalizer
comm
conda
conda-anaconda-telemetry
conda-anaconda-tos
conda-content-trust
conda-libmamba-solver
conda-package-handling
conda_package_streaming
contourpy==1.3.2
cryptography
cycler==0.12.1
debugpy
decorator
distro
exceptiongroup
executing
filelock
fonttools==4.57.0
frozendict
fsspec==2025.3.2
gmpy2
idna
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2
jsonpatch
jsonpointer==2.1
jupyter_client
jupyter_core
kiwisolver==1.4.8
libmambapy
markdown-it-py
MarkupSafe
matplotlib==3.10.1
matplotlib-inline
mdurl
menuinst
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath
nest_asyncio
networkx
numpy==2.2.4
opencv-python==4.11.0.86
packaging
pandas==2.2.3
parso
pexpect
pickleshare
pillow
platformdirs
pluggy
prompt_toolkit
psutil
ptyprocess
pure_eval
py-cpuinfo==9.0.0
pycosat
pycparser
pydantic
pydantic_core
Pygments
pyparsing==3.2.3
PySocks
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML
pyzmq
requests
requests-toolbelt==1.0.0
rich
ruamel.yaml
ruamel.yaml.clib
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
stack_data
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado
tqdm
traitlets
triton==3.1.0
truststore
typing_extensions
tzdata==2025.2
ultralytics==8.3.126
ultralytics-thop==2.0.14
urllib3
wcwidth
wheel==0.45.1
zipp
zstandard
